+++ recpack/datasets/recsys_challenge.py
@@ -3,7 +3,7 @@
 import numpy as np
 import pandas as pd
 from typing import List
-from recpack.data.datasets.base import Dataset
+from recpack.datasets.base import Dataset
 
 from recpack.preprocessing.filters import (
     Filter,