+++ recpack/tests/test_algorithms/test_stan.py
@@ -4,7 +4,7 @@ import pytest
 from scipy.sparse import csr_matrix
 
 from recpack.algorithms.stan import STAN
-from recpack.data.matrix import InteractionMatrix
+from recpack.matrix import InteractionMatrix
 
 
 @pytest.fixture()
@@ -26,9 +26,7 @@ def mini_dataset():
             "tst": [1, 2, 3, 1, 2, 4, 5, 6, 7],
         }
     )
-    return InteractionMatrix(
-        df, user_ix="uid", item_ix="iid", timestamp_ix="tst"
-    )
+    return InteractionMatrix(df, user_ix="uid", item_ix="iid", timestamp_ix="tst")
 
 
 @pytest.fixture()
@@ -42,38 +40,27 @@ def mini_test_dataset(mini_dataset):
 
 
 def test_init(algo):
-    assert (
-        algo.identifier == "STAN(K=2,distance_from_match_decay=1.0,"
-        "interaction_decay=1.0,session_decay=0.2)"
-    )
+    assert algo.identifier == "STAN(K=2,distance_from_match_decay=1.0," "interaction_decay=1.0,session_decay=0.2)"
 
 
 def test_fit(algo, mini_training_dataset):
 
     algo.fit(mini_training_dataset)
 
-    expected_interaction_positions = np.array(
-        [[1, 2, 3, 0], [0, 0, 1, 2], [0, 0, 0, 0], [0, 0, 0, 0]]
-    )
+    expected_interaction_positions = np.array([[1, 2, 3, 0], [0, 0, 1, 2], [0, 0, 0, 0], [0, 0, 0, 0]])
     np.testing.assert_array_equal(
         algo.session_interactions_positions_.toarray(),
         expected_interaction_positions,
     )
 
     expected_session_timestamps = [[3, 2, 0, 0]]
-    np.testing.assert_array_equal(
-        algo.historical_session_timestamps_.A.T, expected_session_timestamps
-    )
+    np.testing.assert_array_equal(algo.historical_session_timestamps_.A.T, expected_session_timestamps)
 
 
-def test_compute_session_similarity(
-    algo, mini_training_dataset, mini_test_dataset
-):
+def test_compute_session_similarity(algo, mini_training_dataset, mini_test_dataset):
     algo.fit(mini_training_dataset)
 
-    sim = algo._compute_session_similarity(
-        mini_test_dataset.last_timestamps_matrix
-    )
+    sim = algo._compute_session_similarity(mini_test_dataset.last_timestamps_matrix)
     expected_similarity = [
         [0, 0, 0, 0],
         [0, 0, 0, 0],
@@ -88,18 +75,12 @@ def test_compute_session_similarity(
     np.testing.assert_array_almost_equal(sim.toarray(), expected_similarity)
 
 
-def test_compute_session_weights(
-    algo, mini_training_dataset, mini_test_dataset
-):
+def test_compute_session_weights(algo, mini_training_dataset, mini_test_dataset):
     algo.fit(mini_training_dataset)
 
-    sim = algo._compute_session_similarity(
-        mini_test_dataset.last_timestamps_matrix
-    )
+    sim = algo._compute_session_similarity(mini_test_dataset.last_timestamps_matrix)
 
-    session_weights = algo._compute_session_similarity_weights(
-        mini_test_dataset.last_timestamps_matrix, sim
-    )
+    session_weights = algo._compute_session_similarity_weights(mini_test_dataset.last_timestamps_matrix, sim)
 
     expected_weights = [
         [0, 0, 0, 0],
@@ -108,14 +89,10 @@ def test_compute_session_weights(
         [np.exp(-4 / 5), np.exp(-5 / 5), 0, 0],
     ]
 
-    np.testing.assert_array_almost_equal(
-        session_weights.toarray(), expected_weights
-    )
+    np.testing.assert_array_almost_equal(session_weights.toarray(), expected_weights)
 
 
-def test_compute_prediction_scores(
-    algo, mini_training_dataset, mini_test_dataset
-):
+def test_compute_prediction_scores(algo, mini_training_dataset, mini_test_dataset):
     algo.fit(mini_training_dataset)
 
     # fmt: off
@@ -128,9 +105,7 @@ def test_compute_prediction_scores(
         ]
     )
     # fmt: on
-    predictions = algo._compute_prediction_scores(
-        session_similarities, mini_test_dataset
-    )
+    predictions = algo._compute_prediction_scores(session_similarities, mini_test_dataset)
 
     # fmt: off
     expected_predictions = [
@@ -144,9 +119,7 @@ def test_compute_prediction_scores(
     np.testing.assert_array_equal(predictions.toarray(), expected_predictions)
 
 
-def test_compute_prediction_scores_single_neighbour(
-    algo, mini_training_dataset, mini_test_dataset
-):
+def test_compute_prediction_scores_single_neighbour(algo, mini_training_dataset, mini_test_dataset):
     algo.fit(mini_training_dataset)
 
     # fmt: off
@@ -159,9 +132,7 @@ def test_compute_prediction_scores_single_neighbour(
         ]
     )
     # fmt: on
-    predictions = algo._compute_prediction_scores(
-        session_similarities, mini_test_dataset
-    )
+    predictions = algo._compute_prediction_scores(session_similarities, mini_test_dataset)
 
     # fmt: off
     expected_predictions = [
@@ -199,9 +170,7 @@ def test_predict(algo, mini_training_dataset, mini_test_dataset):
     ]
     # fmt: on
 
-    np.testing.assert_array_almost_equal(
-        predictions.toarray(), expected_predictions
-    )
+    np.testing.assert_array_almost_equal(predictions.toarray(), expected_predictions)
 
 
 def test_fit_no_interaction_matrix(algo, mat):