+++ recpack/tests/test_algorithms/test_wmf.py
@@ -3,16 +3,14 @@ import numpy as np
 from scipy.sparse import csr_matrix
 
 from recpack.algorithms import WeightedMatrixFactorization
-from recpack.data.matrix import to_binary
+from recpack.matrix import to_binary
 
 
-@pytest.mark.parametrize("num_users", [4, 5, 10])#, 4, 5, 10])
+@pytest.mark.parametrize("num_users", [4, 5, 10])  # , 4, 5, 10])
 def test_wmf_different_shapes(num_users):
     # Regression Test: Used to throw an error with some sizes
 
-    wmf = WeightedMatrixFactorization(
-        confidence_scheme="minimal", num_components=3, iterations=100
-    )
+    wmf = WeightedMatrixFactorization(confidence_scheme="minimal", num_components=3, iterations=100)
 
     values = np.random.randint(0, 2, size=20)
     users = np.random.randint(0, num_users, size=20)
@@ -27,9 +25,7 @@ def test_wmf_different_shapes(num_users):
 
 @pytest.mark.parametrize("cs", ["log-scaling", "minimal"])
 def test_wmf(cs):
-    wmf = WeightedMatrixFactorization(
-        confidence_scheme=cs, num_components=3, iterations=200
-    )
+    wmf = WeightedMatrixFactorization(confidence_scheme=cs, num_components=3, iterations=200)
 
     values = [2, 5, 4, 1, 3, 4, 3]
     users = [0, 0, 1, 1, 2, 2, 2]
@@ -51,20 +47,14 @@ def test_wmf(cs):
     values_pred = [1, 2, 3]
     users_pred = [1, 0, 1]
     items_pred = [0, 0, 1]
-    pred_matrix = csr_matrix(
-        (values_pred, (users_pred, items_pred)), shape=test_matrix.shape
-    )
+    pred_matrix = csr_matrix((values_pred, (users_pred, items_pred)), shape=test_matrix.shape)
     prediction = wmf.predict(pred_matrix)
 
     exp_values = [1, 0, 0, 1, 1, 0]
     exp_users = [0, 0, 0, 1, 1, 1]
     exp_items = [0, 1, 2, 0, 1, 2]
-    expected_prediction = csr_matrix(
-        (exp_values, (exp_users, exp_items)), shape=test_matrix.shape
-    )
-    np.testing.assert_almost_equal(
-        prediction.toarray(), expected_prediction.toarray(), decimal=1
-    )
+    expected_prediction = csr_matrix((exp_values, (exp_users, exp_items)), shape=test_matrix.shape)
+    np.testing.assert_almost_equal(prediction.toarray(), expected_prediction.toarray(), decimal=1)
 
 
 def test_wmf_invalid_confidence_scheme():